Add support for secondary Android N roms (>= 7.0)
In Android 7.0 a change was introduced in the init system that prevents
services from being started if no SELinux is defined for them. This
caused Android roms whose version is greater than or equal to 7.0 not to
boot as secondary roms. This is caused by the init system setting the
wrong SELinux file contexts for the secondary system partition since it
resets the SELinux contexts for all files on the data partition,
including all files of secondary roms.

To prevent this, we comment out the problematic line in the init.rc from
the secondary boot image. Further, we now use the correct file_contexts
file from the secondary boot image as there was a name change in the
transition to M (additional '.bin').

Change-Id: Ifd054e25136ebed4f3d67e0f9347e0949a9279c0
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>